feat: Write up a simple prototype for the project
This is just an experiment for the algorithm, it may not be optimal but it depends on the complexity of any single namespace.

We should probably write the real thing in Go because of the nice Level DB and Consul integration. Though, parsing the JSON and building dicts is going to be a pain.