feat: Implement creating documents
When a user creates a document, he becomes the owner of said document. The ACL directive is sent to Lesotho; however, the `user` field, which in this case is the ID in our "json database", cannot be sent directly. This is because a string that _can_ be parsed as an int _will_ be parsed as an int. To circumvent this, the ACL stores user IDs prefixed with a "u".

Users can edit and read documents. Neither of these functionalities is implemented so far.
For simplicity, the CLI prints ALL documents on the server and the user can select one of them.
If you do not have the required permissions for the selected document, you won't be able to edit/read the document.